+++ validators/unconnectedTrafficLights/index.js
